Add JsonException
Change SendyException into an interface to allow extending native
exceptions like JsonException.
