+++ src/ApiException.php
@@ -3,33 +3,9 @@
 namespace Sendy\Api;
 
 /**
- * @deprecated This class exists for backwards compatibility and may be removed in a future version.
- * @todo Replace internal usages with the new more granular exceptions in `Sendy\Api\Exceptions`.
+ * @deprecated This interface exists for backwards compatibility and may be removed in a future version.
  * @internal
  */
-class ApiException extends \Exception
+interface ApiException extends \Throwable
 {
-    /** @var array<string, string[]> */
-    private array $errors = [];
-
-    /**
-     * @param string $message
-     * @param int $code
-     * @param \Throwable|null $previous
-     * @param string[][] $errors
-     */
-    public function __construct(string $message = "", int $code = 0, ?\Throwable $previous = null, array $errors = [])
-    {
-        $this->errors = $errors;
-
-        parent::__construct($message, $code, $previous);
-    }
-
-    /**
-     * @return array<string, string[]>
-     */
-    public function getErrors(): array
-    {
-        return $this->errors;
-    }
 }
