+++ src/Exceptions/SendyException.php
@@ -4,6 +4,12 @@ namespace Sendy\Api\Exceptions;
 
 use Sendy\Api\ApiException;
 
-abstract class SendyException extends ApiException
+interface SendyException extends ApiException
 {
+    /**
+     * Get the error details from the response.
+     *
+     * @return array<string, string[]>
+     */
+    public function getErrors(): array;
 }
