+++ src/Exceptions/ServerException.php
@@ -5,6 +5,7 @@ namespace Sendy\Api\Exceptions;
 /**
  * Represents an HTTP 5xx error that occurred during a request to the Sendy API.
  */
-class ServerException extends SendyException
+class ServerException extends \Exception implements SendyException
 {
+    use HasErrors;
 }
