+++ src/Exceptions/ValidationException.php
@@ -2,6 +2,9 @@
 
 namespace Sendy\Api\Exceptions;
 
+/**
+ * Represents an HTTP 422 error that occurred during a request to the Sendy API.
+ */
 class ValidationException extends ClientException
 {
 }
